add a new rule for build_info.f90, so that we are sure that it is regenerated any time we recompile.  Previously, the script was not rerun when a file changed, so the build info could be out of date.
Now there is a new class of files, sf90sources, that do not get
the dependency check done to them.  Rightnow, build_info.f90 is the
only such example of this type of file.  This allows it to have as
dependencies all of the other files, thus guarranting that it
gets regenerated.