Now allow us to have StateData that is not stored in the checkpoint.
If store_in_checkpoint = false then the statedata multifabs will not
be written to the checkpoint diretory (though the times, domain, etc
will be written in the header).    We set "nsets" = 0 when we write.

Then when we read, for each statedata we read "nsets" to know whether
to read 0, 1, or 2 multifabs (1 if only new data, 2 if both old and new data).