Some new features.
If one types "touch dump_and_continue" in the run directory, a
checkpoint file and a plot file will be generated and then the run
will continue.  The "dump_and_continue" file will be deleted.

If one types "touch stop_run" in the run directory, the run will stop
without generating a checkpoint.  The "stop_run" file will be deleted.

If one types "touch dump_and_stop" in the run directory, a checkpoint
file and a plot file will be generated and then the run will stop.
The "dump_and_stop" file will be deleted.