CC linear solver:  The convergence test and the way to compute residues have changed.
In the old way, at the beginning of ml_cc, a new RHS was computed
using the initial guess, and then what the solver really solved was
the redefined problem: Lap(\delta \phi) = RHS_new, where RHS_new =
RHS_0 - Lap(\phi_0).  After the solver converged on the new problem,
the full solution is updated as \phi += \delta \phi.

In the new way, the convergence test uses the original RHS and the
latest full solution.