Add the option to set
amr.initial_grid_file =

as well as

amr.regrid_file =

This is similar to the regrid_file option, except that:

1) it is only used to set the initial grids file; there is no file specified
for later regrids.
2) the max_grid_size criterion is not enforced on the initial grids generated
this way, rather it is imposed after the grids are read.  The point is to allow
us to initially read in grids specified by an external routine which did not
know about max_grid_size.