turns out that when we specified an alternate location in the .ini file for a microphysics routine, like:
addToCompileString = NETWORK_TOP_DIR='$ASTRODEV_DIR/networks' NETWORK_DIR=aprox13

it was not using the ASTRODEV_DIR that was set via

extSrcDir = /home/zingale/gfortran-testing/AstroDev/
extSrcCompString = ASTRODEV_DIR

but instead was using whatever the value was through the system.  This
is because we were using os.system() and bash to do the make.  Now we instead
switch all of the build stuff from os.system() to subprocess -- this is a lot
more flexible.  As a result, we now have access to stdout, stderr, and the
return value