ParallelDescriptor: Communicator(), MyProc(), and NProcs() return values in the "local" communicator
A large number of funtions assume the existence of only a single, global
communicator, even though they operate on local objects. This approach worked
fine until we introduced the notion of distinct MPI groups for computation and
in-transit analysis. These MPI functions frequently break when operating on the
in-transit processes because they request the computation communicator, which
is undefined on those processes.  Rather than introducing a million if-else's
to all these functions, forcing them to explicitly return CommunicatorComp() or
CommunicatorInTransit(), we just let Communicator() pick the correct
communicator automatically.