C_BaseLib: glorious re-design of the sidecar interface (again)
This is our glorious new interface to the sidecars. It obviates the need
for abstract base "callback" analysis classes and replaces them with
void pointers to blackbox functions. This allows the user to write
arbitrary analysis workflows, rather than requiring her to define
Initialize(), DoAnalysis(), and Finalize() functions, as was required
with the previous interface.

Marc Day was responsible for most of this work. I just did some
peripheral cleaning afterwards.