C_BaseLib: remove "__LP64__" preprocessor macro from list of macros describing long pointers
This macro does not uniquely describe long pointers. It leads to compile-time
errors on Mira (BG/Q) because many variables in FPC.cpp are duplicately
declared, since both "__ppc__" and "__LP64__" are defined.

This removal should (hopefully) not affect any architectures we run on, since
at least one of the other macros should be defined.