Tutorials: add "beta" to argument list of solve_with_HPGMG() in MultiGrid_C
HPGMG internally calculates beta at cell faces from the cell-centered beta
MultiFab because BoxLib and HPGMG store cell face data in very different ways.
It's much easier to do the averaging by hand than it is to translate array
indices.

However, in order to be consistent with the sign conventions of the other
solvers, we want to calculate grad(Phi) in the same way that they do. So we use
an AbecLaplacian to calculate the gradients. But AbecLaplacian needs the beta
at cell faces, not cell centers. So we add beta to the arguments list.

This also fixes a compile-time error if running this tutorial with HPGMG: the
variable "abec_operator" was never declared.