C_mk: remove "-xHost" flag from Intel compiler flags on Edison and Cori
The Cray compiler wrappers for the Intel compilers already contain the correct
instruction set flags for both architectures ("-xCORE-AVX-I" on Edison and
"-xCORE-AVX2" on Cori). Furthermore, "-xHost" uses the optimal instruction set
on the architecture which hosts the compiler, which may be different than that
of the compute nodes.

This commit partially reverts commit 0088bae.