ParticleContainer: move OpenMP parallel region from grid loop to PBox loop in AssignCellDensitySingleLevel()
If the OpenMP region parallelizes over grids, and if we have a small number of
large grids and a large number of OpenMP threads (which is the typical Nyx use
case), then this routine scales extremely poorly. So instead we parallelize
threads over PBoxes within each grid, which leads to much better thread
scaling.