add an indicator if we are running on a non-default branch + bug fix
now we add an '*' next to the date on the master page if any of the git
repos are not on "suite.default_branch".  This helps delineate the
normal tests from those we run occassionally on feature branches.

Also fix a crash if we ran with only self-tests (no output file
is ever defined)