Add a CUDArena memory management system using CUDA UM
CUDA Unified Memory (available for NVIDIA GPUs of the Kepler generation
or newer) allows a single pointer to point to a region of memory that can
live either on the CPU or the GPU, and switches back and forth on demand.
This is done by declaring a managed memory allocation using cudaMallocManaged,
a drop-in replacement for malloc.

The CUDArena as currently written is a drop-in replacement for BArena.

Credit for the name and original idea belong to Brian Friesen.