Add capability to do UM prefetches to/from device
By setting the new parameter do_device_transfers in MFIter, one can
enable the ability to prefetch any FArrayBox accessed by that MFIter
to the device. Then, when the MFIter index is incremented or the
object is destroyed, all of the prefetches are reversed, bringing
the data back to the host. For now, we default to moving everything
to GPU 0. In the future, we should generalize this so that we can
operate on multiple devices.