Add a device copy of data to BaseFab, with hooks up to StateData
Using the MFInfo struct, a MultiFab can optionally allocate a device copy
of its data. A hook has been added to StateData so that this can be done there
as well. Right now the device implementation is for NVIDIA GPUs with CUDA,
but in principle other devices that require offloading could be used.