Do nothing with BL_TO_FORTRAN_3D when CUDA is enabled
Since we are using managed memory to hold a copy of the box bounds,
we want to directly access the pointer containing that memory, rather
than save a temporary copy. The CUDA version automatically enables
the same functionality as the macro because the lo and hi vectors
contain zeroes in unused spatial dimensions.