Always reset device memory when applying a Box transformation
The original idea was to only reset the shared pointer containing the memory
for routines that returned a new box by value rather than modifying the
existing box by reference. Thus we could keep the CUDA managed memory without
reallocating it if one did Box bx = mfi.validbox(), but create a new allocation
if one did Box bx = surroundingNodes(mfi.validbox()). However, this is defeated
by the use case where one first creates the box copy, and then applies the
transformation operation in a second step. So the only safe approach is to always
reset the allocation if the box is transformed.