Delay initialization of the RealBox device memory
Geometry contains a static RealBox object, and it was creating problems
that a CUDA API call (to allocate memory) was being called before we
got to the initialize_cuda step. By delaying allocation of the device
memory until the first time we actually access the data, we still get
the desired behavior from RealBox while not breaking the intended
behavior that no CUDA API calls are done until we've mapped the MPI ranks
to GPUs.