Add MFIter::get_fortran_pointer to assist with scalar values for device routines
There are cases in the code where we pass scalar values like the current time
to a Fortran kernel, and we pass it by reference. An example is AmrLevel::Derive.
For backward compatibility we do not want to change the interface -- that would break
a lot of codes. But we also want to pass a device pointer for the CUDA version, and
to allow asynchrony across MFIter indices. The solution is to store a shared_ptr with
a pinned host (or device, but I've gone with pinned host for now to avoid a cudaMemcpy call)
copy of the data, so that the lifetime of the data is the lifetime of the MFIter.
For the CPU version of the code, the function effectively does nothing.