Do not use hardcoded numbers for real kinds
Such numbers are compiler specific. In particular, the NAG compiler fails to
compile this line. One must use some platform independent way to get the real
kind. In this case, the number 8 means double precision in gfortran, so I have
changed it to kind(0d0) where 0d0 is a double precision number, and kind()
returns the proper kind on all platforms. Now this line compiles with NAG.