Compare real kinds directly
In order to determine if the real kinds are equal, one can just compare them
directly, they are just integers. This is simpler, and also allows these lines
to be compiled with NAG. For some reason NAG does not seem to have `c_sizeof()`
as part of the `iso_c_binding` module and so it fails on these lines. This
commit fixes it.