GNUMake: remove links to FCVODE, the old Fortran wrapper to CVODE
FCVODE comes with CVODE and provides a Fortran interface to the CVODE library,
but is not thread-safe. Recently the SUNDIALS developers provided us with a
Fortran 2003 interface to CVODE which uses iso_c_binding to most of the
functions in the library. So remove the links to the old FCVODE library since
we don't need it anymore.