Create a dimension agnostic, multi-component filcc
This version of filcc is both dimension agnostic, and works over an arbitrary
number of components. It gives the same results as the dimensionful versions of
filcc. It will be necessary for performance reasons to use a version like this
on the GPU.

This also applies the new filcc to amrex_fab_filcc.