EB, Combustor: add function get_neighbor_cells_int_single()
This function behaves similarly to get_neighbor_cells_int(), except it returns
a single value for the requested neighbor cell, rather than returning a 3x3x3
array containing values for all neighbors. This single-value approach is
friendlier to function inlining and vectorization, and also yields better
memory access patterns because a compiler attempting to vectorize a loop which
calls this function need not generate private 3x3x3 arrays for each lane of a
vector processing unit.