Enable coarse and fine level selection in fextract
fextract previously would extract data on all levels by default. This adds
options to select a different set of levels; for example, a use case might
be to only see the data on the finest level so you can examine where the
refined regions are. The new options are -c/--coarse_level and -f/--fine_level.