Tools: change XL MPI compiler wrappers on summitdev to "mpifort"
The OpenMPI man page for "mpif90" and "mpif77" notes that these wrappers
are deprecated, and that one should instead use "mpifort" for all
Fortran files, both free and fixed format. The observant user will note
also that all three wrappers point to the same file ("opal_wrapper").

This change is currently in contradiction with the documentation on the
OLCF summitdev website, which recommends continued use of "mpif77" and
"mpif90".