Update XL flags (#186)
* Tools: move common XL Fortran compiler flags to GENERIC_IBM_FLAGS

* Tools: add flag to enforce fixed form format in XL FORTRAN 77 compiler

This compiler makes assumptions about the source code format based on
file suffixes, which leads to errors when compiling some AMReX files.
Fix this by explicitly stating the source code style with this flag.

* Tools: change XL MPI compiler wrappers on summitdev to "mpifort"

The OpenMPI man page for "mpif90" and "mpif77" notes that these wrappers
are deprecated, and that one should instead use "mpifort" for all
Fortran files, both free and fixed format. The observant user will note
also that all three wrappers point to the same file ("opal_wrapper").

This change is currently in contradiction with the documentation on the
OLCF summitdev website, which recommends continued use of "mpif77" and
"mpif90".