Do not initialize dt_level on levels that do not yet exist
This initialization is problematic if the level 0 timestep when the simulation
starts is significantly different from the timestep later on in the simulation
when we add refinement. In particular, this will happen if the simulation code
shrinks the initial timestep by some factor. If the fine levels are only added
later on once the timestep has caught back up to its normal limited level, then
the first timestep on the fine level will be effectively limited by the initial
level 0 timestep, which can sharply decrease the timestep for no reason.