Fix a potential numerical overflow in BoxArray::minimalBox
If the number of points per dimension N is such that N^3 is larger than
a 32-bit integer (or N^2, for 2D), then this implementation of minimalBox
will overflow when accumulating into an n_pts variable. Solved by replacing
the value holding the number of points with a long, which is consistent with
similar calculations elsewhere in BoxArray.

Fixes #357