Fill in last_checkpoint/last_plotfile after a restart
Castro and Nyx have logic in main.cpp that says to only dump a checkpoint
if the coarse timestep of the last checkpoint is older than the current
coarse timestep. In the edge case where the simulation doesn't actually
take any timesteps after a restart, this results in a checkpoint and plotfile
being written from the same data we restarted from, which unnecessarily clobbers
the old data and creates a duplicate checkpoint/plotfile.

If the user explicitly wants a new checkpoint on restart, perhaps because
they are doing some kind of transformation, they can still take advantage
of the flags checkpoint_on_restart and plotfile_on_restart.