Add floating point awareness to the plot_per/check_per logic
The current logic can be slightly erroneous due to floating point issues. For example,
if the timestep is 0.01 and the plot_per is 0.1, you would expect plotfiles at
steps that are multiples of 10, but due to roundoff errors the current logic will sometimes
plot in the 11th step instead of the 10th. By adding a small number to both sides of the
comparison, we can guarantee plotting on exactly the timestep we intended.