Add optional flag to RegridOnly for disabling I/O
RegridOnly is useful as a public interface to regridding for the user
code to call. For cases where the user wants to call it during a
simulation, rather than immediately after restart, the I/O options
aren't meaningful, so this allows the user to disable that in their
call to RegridOnly (while still potentially allowing it in other
cases where plotfile_on_restart et al. are desired).