Allows to create FPInfo objects with an associated EB2::IndexSpace.
For enabled EB these changes allow to choose the EB2::IndexSpace which
will be used for the interpolation regions in FillPatchTwoLevels.

This commit also makes it possible to use FillPatchTwoLevels without any
EB2::IndexSpace built at a prior stage. In this case it will do the same
communication logic as if AMReX would not be built with enabled EB.