openPMD-api: v0.9.0 provides .pc files
In order to build with openPMD-api, we can rely on the installed
`pkg-config` (`.pc`) file provided in openPMD-api 0.9.0 or newer.

We also set an RPATH in the WarpX executable to avoid having to
load the shared depending library via `LD_LIBRARY_PATH` at runtime.
Further dependencies, if provided by spack, have an RPATH set by
default as well.