This commit fixes a strange issue with AMReX cell bilinear interp.
We setup a 1D simulation (planar flame) on a 2D grid.  After some time steps there would be significant transverse velocities (which should be zero) that affected the flame.  This only occured with the Intel compiler with DEBUG=False and with the cell_bilinear interpolation.  Compiling with GNU orIntel with DEBUG=True did not produce this behavior.  Rearranging the interpolation procedure to the method typically done by students when interpolating tables by hand seems to work fine where the data is first interpolated long x-lines, then the x-line interpolated data is interpolated in the y-direction, and finally the y-interpolated data is interpolated in the z-direction to give the final value.