Headers: No Public Scope Using Namespace
Writing `using namespace XY;` in public scope of a header file
pollutes downstream namespaces that include that file directly
or even indirectly.

This can be cumbersome and lead to nifty naming conflicts that
cause unclear selection of classes or ADL lookup issues. So
one just avoids this as good practice and we did not do it in
many places anyway.