Avoid race conditions in reduction code called multiple times
cuda-memcheck --tool racecheck reported race conditions in this code
if we called the block reduction multiple times in a kernel, since some
warps in the first reduction could be reading from shared memory at the
same time as other warps in the second reduction were writing to it. By
inserting a syncthreads prior to the first write to shared memory, we
avoid this (at some performance cost for the cases where we only call
the reduction once in a kernel).