Fix Segfault: Create Particle Functor
Copy over the full create-particle functor including its
data via managed memory. Previously, we used a host-side
vector to initialize the functors and then did not copy over
the the state to device.

As always, this does not show up as an error on Summit but as
a slowdown at runtime. This crashed with segfaults on all other
GPU machines.