CXXSTD: Project Control of C++ Std
Add a new option, defaulting to C++11, to give projects control over
the required C++ standard of a project.

This improves the current logic, which does not cleanly fail if a
compiler does not support e.g. C++14 but instead gave hard-to-read
error messages such as "hm, dunno what std::make_unique is" :)
Now the compiler will fail with messages such as:
"gcc: error: unrecognized command line option ‘-std=c++14’"
which is better and safes us from re-implementing a ton of logic
that CMake provides out-of-the-box.