CMake: AMReX::amrex Alias
Adding a project alias on `CMakeLists.txt` level that is
identical to the installed CMake namespace.

This is a community-convention in CMake and makes sure that
both `add_subdirectory(AMReX_DIR)` as well as `find_package(AMReX)`
targets can just be accessed as `AMReX::amrex` in all cases in
user code.