Fix: read_raw_data Tool for Reduced (#879)
The read_raw_data python scripts were not consistently returning
arrays.
This fixes it, making sure that both a reduced diagnostics that has
only been run for one step (e.g. in a test case) as well as for
arbitrary many steps returns a dict of arrays.

Seen with BeamRelevant reduced diagnostics.