Async out (#845)
* AsyncOut using a background thread.  This is based on our previous
implementations of async output.

In our first implementation, we use std::async to launch a thread for
the output job.  There are two issues with that first implementation.
First, it does not scale to a large number of MPI processes.  We do
not want every MPI process to write to disk simultaneously.  To avoid
that, the first implementation writes a small file to the disk and
uses that to coordinate.  This way of message passing using file does
not scale.  The second issue is it is hard to manage multiple jobs.
Ideally we would like the second job to start after the first job is
finish to avoid launching too many threads writing to the disk.

In our second implementation, MPI is initialized with support for
multiple threads.  This solves the first issue above.  However, the
second issue still exists.  To write a plotfile with multiple
MultiFabs like the functions in AMReX_PlotFileUtil, the second
implementation has a new function to manually merge multiple
MultiFabs' data into one contiguous chunk of memory for a std::async
job.  However this cannot be easily extended to AmrLevel's
WritePlotfile function.

In this third implementation, we launch a background thread
specifically for async output.  The thread persists during the run
waiting on a C++ conditional variable.  The thread is woke up when a
job is submitted to the async job queue.  As in the second
implementation, MPI_THREAD_MULTIPLE is used.  The use of a background
thread and a job queue provides flexibility.  For example, writing
plotfile in AmrLevel requires only a few lines of changes.  One can
submit multiple jobs to the queue, as long as there is enough memory,
without worrying about launching too many threads.