CMake: Fix CUDA Host Check
The check did only write a regular message due to a typo of the
message type.

Then checking the test, I realized it is not save:
- false negatives are raised if e.g. CXX is set to g++ while
  CUDAHOSTCXX defaults to c++
- checking only env variables is not sufficient to validate user
  intent: setting explicitly off compilers with -DCMAKE_...
  variables slips through

The only save check we can implement is checking compiler name
and ID, but that one is blocked by a missing upstream feature:
  https://gitlab.kitware.com/cmake/cmake/-/issues/20901

All in all: fallback now to (flawed) env-logic but only throw a
warning.