Optimization of TagBoxArray::collate
Use a new way to remove duplicated tags.  Instead of removing duplicated from vector, we remove
duplicates when tags are still stored in FabArray.  Because they are in FabArray with a BoxArray
that has no overlapped valid cells, we can use ParallelAdd to remove duplicates due to the overlap
of valid cells with ghost cells.

Assertion in AmrCore on blocking factor to make sure that there are no overlapped valid cells in
TagBoxArray after coarsening.