Optimization of regrid
New BoxList::maxSize function that keeps Boxes after chopping in an order that is easy to merge them
back into original Box.

New BoxList::order_simplify function that assumes the BoxList is in the order that we only need to
search the next Box in the list.

Minor optimization of BoxList::intersect.

Cache simplified BoxList inside BoxArray.