Communication buffers (#1186)
Use The_Cpu_Arena (BArena) instead of The_Pinned_Arena (CArena) for communication buffers to avoid
the CArena fragmentation issue with the type of problems that allocate increasingly larger memory in
a contiguous chunk during simulation.  CArena then cannot effective use previous allocation chunks
because they are slightly smaller.  The fix is for CPU build only.  We still need to have a strategy
for this for GPU build.