EB outside domain (#1201)
* EB outside domain

Add a runtime parameter `eb2.extend_domain_face` to control EB generation outside the domain.  The
default is false.  That is the implicit function will be used for cells outside the domain.  This
preserves the current behavior.  If it is set to true, the geometry information is then extrapolated
from domain faces.

* fix fillFab for EB_levelset