The average_down_faces calls between AMR levels in the linear solvers need to see periodicity (#1221)
The average_down_faces calls between AMR levels need to use the interfaces that sees the periodicity.  This requires adding that functionality to EB_average_down_faces (it has already been added to non-EB average_down_faces.)

## Summary
Recently we added the functionality for an average_down_faces call to average fine data from one side to average not only onto the face under it but also onto the face on the other side of the domain if periodic in that direction.  This PR 1) extends that functionality to the EB_average_down_faces call and 2) uses this new functionality when averaging down face-based coefficients between AMR levels in the cell-centered linear solvers.

## Checklist
 
The proposed changes:
- [X] fix a bug or incorrect behavior in AMReX
- [X] add new capabilities to AMReX
- [X] changes answers in the test suite to more than roundoff level
- [X] are likely to significantly affect the results of downstream AMReX users
- [ ] are described in the proposed changes to the AMReX documentation, if appropriate