Optimization of regrid (#1199)
## Summary

New BoxList::maxSize function that keeps Boxes after chopping in an order that is easy to merge them
back into original Box.

New BoxList::order_simplify function that assumes the BoxList is in the order that we only need to
search the next Box in the list.

Cache simplified BoxList inside BoxArray.

Use simplified BoxArray in FPinfo.

This commit will make most regression tests fail because it changes the grids.  But if the comparison is
done with `fcompare....ex -a`, it should only change answers at roundoff level unless the test is very
sensitive.