oneAPI Beta 8 (#1243)
In beta8, C++11 user defined literals are broken because of the use of long
double.  We have implemented a workaround by overloading operator "" _rt
with `const char*` argument and our own `atof`.  But, because Clang does not
evaluate the new version of the operator at compile time, we now avoid using
`_rt` in AMReX.  In the long term, we may have to remove it entirely if the
long double issue is not resolved in DPC++.

With the workaround, we can enable the DPC++ CI test again.