Domain precision (#1209)
Due to issues with pointing point precision, there are points inside the `ProbDomain()` that do not map to cells inside the `Domain()` box when binned using something like the `getParticleCell` function. This PR adds an additional domain, the "RoundoffDomain()", to the Geometry object. Points inside the roundoff domain are always sure to map to cells inside `Domain()`. The bounds are calculated when the Geometry object is constructed using bisection. 

This helps solve a number of issues that can happen when you have particles very close to the domain boundaries.

While I have tested this with Nyx and WarpX, there is a chance that there could be a downstream effect on other codes. If the bisection fails when setting the roundoff domain on some problem, then we may need to adjust some tolerances.