Tensor solver and extension of EB domain face
ParmParse parameter eb2.extend_domain_face is now true by default.  The
reason for the change is that for most applications it's hard to imagine one
would want it to be false.  It will also solve the EB tensor solver issue
when there is a cut cell just outside the domain abutting a covered valid
cell.

Recent changes to the tensor solvers are incorrect and are reverted here.
They are incorrect because the ghost cells in the modified functions
actually have values at the cell centered.  Although the users put domain
face values in the ghost cells, that is not what the solver does.  The
solver stores the boundary values in boundary registers.  Before the stencil
is being applied, bc function is called to fill the ghost cells with
properly interpolated or extrapolated values so that the stencil operations
are just like working on normal interior cells.

Revert "Correcting tensor cross terms computation for periodic bcs (#1254)"
This reverts commit 1197adc5be6144bae9e362c114ca16b879006180.

Revert "Changing computation at inflow/outflow for tensor solve (#1235)"
This reverts commit b391885e918d63fa1741f954d2792e31426a6204.

Revert "Changing computation at outflow for the EBTensor (#1187)"
This reverts commit 9cd538802ea5b3ab7e6db10d6046933e937aa449.