Remove HOST and HOSTNAME from make system
PR #1275 fixed a bug in the use of HOST and HOSTNAME in the make system.
However, the fix broke on Summit because HOSTNAME on Summit login nodes is
set to something like `login1`.  We now remove the use of HOST and HOSTNAME
variables and rely on `hostname -f` and `hostname -s` commands, which is the
behavior before the fix in #1275.