Modify the use of HOST and HOSTNAME in make system (#1303)
## Summary

PR #1275 fixed a bug in the use of `HOST` and `HOSTNAME` in the make system.
However, the fix broke on Summit because `HOSTNAME` on Summit login nodes
is set to something like `login1`.  We now set `host_name` to `hostname -f` and
`host_name_short` to `hostname -s` command, which is the behavior before the fix
in #1275.  We add a new make variable `host_name_env` that is set to
`HOSTNAME` and `HOST` from the environment and use it for tulip.

## Additional background

## Checklist

The proposed changes:
- [ ] fix a bug or incorrect behavior in AMReX
- [ ] add new capabilities to AMReX
- [ ] changes answers in the test suite to more than roundoff level
- [ ] are likely to significantly affect the results of downstream AMReX users
- [ ] are described in the proposed changes to the AMReX documentation, if appropriate