Fix tensor solver with periodic boundaries (#1383)
## Summary

Remove erroneous calls to EnforcePeriodicity.  When coarse/fine boundary
meets periodic boundary with fine grids form an L-shape, the uncovered fine
ghost cell at the corner is actually "multi-valued". That is the cells with
the same indices has different values in different boxes.  That's why
calling EnforcePeriodicity is erroneous.  For CPU runs, this bug introduces
deterministic errors, but the solver still converges.  For GPU runs, this
causes convergence issues too because EnforcePeriodicity is not
deterministic when the values inside domain are not consistent.

## Additional background

## Checklist

The proposed changes:
- [x] fix a bug or incorrect behavior in AMReX
- [ ] add new capabilities to AMReX
- [x] changes answers in the test suite to more than roundoff level
- [x] are likely to significantly affect the results of downstream AMReX users
- [ ] are described in the proposed changes to the AMReX documentation, if appropriate