Random Number Generation on Device (#1363)
In order to support RNG on device with DPC++, we have to change the API
because DPC++ does not support global device variables.  We introduce a new
ParallelForRNG function that takes a device lambda with an extra
RandomEngine argument.  Random numbers can be generated by calling
amrex::Random(RandomEngine const&).  To maintain backward compatibility in
the short term, the old way of calling `amrex::Random()` on device is kept.
But it will be removed soon because no locking is needed in the new way
unlike in `Random()`.  Note that only the device version of
`amrex::Random()` is deprecated.