CMake 3.18+: CUDA Arch Policy (OLD)
Keep the old policy for now to avoid setting the code generation
flags twice.

We should be able to transition this well, but I need to first
find out how device LTO generation flags are handled here, if
at all, and how to restore the default-detection of the local GPU
architecture, e.g. on Summits head-nodes.