make clean & make cleanconfig
`make clean` now does what `make realclean` does.  `make realclean` is kept.
`make cleanconfig` is introduced to do what `make clean` used to do.  For
most users, `make clean` is the one that should be used.

Also add some tests in makefiles to avoid error messages in case when `make
clean` is run on a makefile with the default not suitable for the system.