make clean & make cleanconfig (#1520)
## Summary
`make clean` now does what `make realclean` does.  `make realclean` is kept.
`make cleanconfig` is introduced to do what `make clean` used to do.  For
most users, `make clean` is the one that should be used.

Also add some tests in makefiles to avoid error messages in case when `make
clean` is run on a makefile with the default not suitable for the system.

## Checklist

The proposed changes:
- [ ] fix a bug or incorrect behavior in AMReX
- [ ] add new capabilities to AMReX
- [ ] changes answers in the test suite to more than roundoff level
- [ ] are likely to significantly affect the results of downstream AMReX users
- [ ] are described in the proposed changes to the AMReX documentation, if appropriate