AMReX_Config.H
Based on discussions with @sayerhs, @ax3l, and @mic84, We have decided to
make the following changes to the build systems.

All headers in Src/ will include AMReX_Config.H, which will be generated by
the build systems (both gnu make and cmake).  The file will include all the
macros we currently pass as compiler options.

The purpose of this is to solve some inconsistency issues in the current
systems.  We support both super build and using AMReX as a library.
Moreover, we do not want to force libamrex users to use the same build tool
that is used to build the library.  In this PR, we are put all the macro
definitions in AMReX_Config.H that is included by every AMReX header.