Correct extended host/device lambda internal/no linkage error
MSVC + CUDA has restrictions on internal functions launching
host+device or just device lambda's. We move the lambda's
to be launched by inline functions to work around this issue.

We have to drop the `noexcept` from these lambda's as well
as it caused compiler reflection problems on MSVC when
trying to figure out how to capture the device version