EB2::PolyIF (#1567)
The existing PolynomialIF contains a dynamic size vector.  This makes it
hard to support it on GPU.  We have used an undocumented feature to get this
run on GPU.  However, this code can no longer compile, because the GNU make
system now makes cross execution space calls an error.  So we disable GPU
support for PolynomialIF and add a new PolyIF class with fixed size array
and therefore GPU support.