Eb from stl (#1592)
## Summary
This is a work-in-progress PR that adds a preliminary set of tools and helper functions for using triangulated STL files in amrex based applications. Coupling with AMReX's EB framework will be added in later pull requests. This PR is done early so as to provide a software framework for incremental changes. 

As of now, I have added one helper function that provides a nodal multifab with a marker value indicating inside or outside the STL geometry. I guess it can be used for particle-wall collisions. 

I have also added a test case in the EB folder that generates a nodal multifab with the blanked values for an airfoil.


The proposed changes:
- [ ] fix a bug or incorrect behavior in AMReX
- [x] add new capabilities to AMReX
- [ ] changes answers in the test suite to more than roundoff level
- [ ] are likely to significantly affect the results of downstream AMReX users
- [ ] are described in the proposed changes to the AMReX documentation, if appropriate


Co-authored-by: Hariswaran Sitaraman <hsitaram@el2.ib0.cm.hpc.nrel.gov>
Co-authored-by: Hariswaran Sitaraman <hsitaram@el1.ib0.cm.hpc.nrel.gov>