GNU Make: Fix the library flags
Previously for GPU build, the GNU Make system was very sensitive on which
variable to use for adding extra `libraries`.  Before a certain location,
`LIBRARIES` should be used, but after that `libraries` should be used.  This
is very error prone.  This PR tries to fix this.  Now the user should be
able to use `LIBRARIES` anywhere.  The use of `libraries` should be avoided
because it is still the case that it can only be used after a certain
location in the make file.  Nevertheless, this PR is backward compatible and
it does not force the use of `LIBRARIES`.